fix: parse dpkg progress percent as double, not int
QJsonValue::toInt() returns 0 for non-integer doubles; apt emits
the progress percent as std::fixed decimals (e.g. 11.1111), so every
pmstatus event was parsed as 0 and the progress bar never moved
during multi-package installs (single-package worked because apt
happened to emit integer values). Use toDouble() and truncate
instead, in both onStatusChanged and findEventPercent.

+++ src/amoupdates.cpp
@@ -493,7 +493,7 @@ void AmoUpdates::onStatusChanged(const QString &statusJson)
     // Progress percentage if present.
     const QJsonValue percentValue = obj.value(QStringLiteral("percent"));
     if (percentValue.isDouble()) {
-        const int dpkgPercent = qBound(0, percentValue.toInt(), 100);
+        const int dpkgPercent = qBound(0, static_cast<int>(percentValue.toDouble()), 100);
         if (m_hasGlobalDownloadProgress) {
             setPercentage(50 + dpkgPercent / 2);
         } else {
@@ -571,7 +571,7 @@ int AmoUpdates::findEventPercent(const QJsonObject &obj) const
     if (obj.contains(QStringLiteral("percent"))) {
         const QJsonValue p = obj.value(QStringLiteral("percent"));
         if (p.isDouble()) {
-            return p.toInt();
+            return static_cast<int>(p.toDouble());
         }
     }
     for (auto it = obj.begin(); it != obj.end(); ++it) {
